Bug while detecting non answering battery (#383)
* Bug while detecting non answering battery

Stumbled about this while testing:

`error_count` is local to publish_battery() and there reset to 0 on every call. Therefore it will never reach the later values of 10 or 60 that indicate a problem. Possible solution: make error_count a class variable

* Update dbushelper.py

error_count was still reset on every call regardless of success

Co-authored-by: Louis Van Der Walt <26077665+Louisvdw@users.noreply.github.com>