Add dynamic voltage adaption.
To explain the algorithm, let's assume the configured maximum cell voltage
is 3.5V. If the top cell is at 3.4V and all others are at 3.39 or so,
the pack is well-balanced and we can (most likely) safely charge until
they're all at or slightly below 3.5V.

However, if all other cells are at 3.3 or so, we can only increase the
total pack voltage by 0.1V before the top cell's voltage exceeds 3.5V.
In this case the other cell voltages will not move much, if at all,
due to the highly nonlinear charge curve of LiFePo4 cells.

This method this reliably prevents excessive cell voltages.

The same in reverse applies to the minimum overall voltage when discharging.
The knee in the discharge curve is not as pronounced, but the effect
still holds.