Introduce PENALTY_BUFFER for Overcharge Prevention
This commit introduces a new constant, PENALTY_BUFFER, to make the code more readable and maintainable. PENALTY_BUFFER is added to the penaltySum calculation when a cell's voltage exceeds MAX_CELL_VOLTAGE, thus creating a buffer or safety margin. This results in a stricter overcharge prevention mechanism as it reduces the control voltage further when a cell's voltage exceeds MAX_CELL_VOLTAGE. The value of PENALTY_BUFFER is currently set to 0.010.